requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
djoser==2.3.1

gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
django-cors-headers==4.4.0

firebase-admin==6.6.0
Pillow==11.0.0
python-dotenv==1.0.1
